# API-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebShop24",
    "description": "This is the best WebShop you can find.\n",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://petstore.swagger.io/v2"
    }
  ],
...
  "paths": {
    "/buyer": {
      "post": {
        "tags": [
          "buyer"
        ],
        "summary": "Create a new buyer (registration)",
        "operationId": "createBuyer",
        "requestBody": {
          "$ref": "#/components/requestBodies/Buyer"
        },
        "responses": {
          "400": {
            "description": "User Data incomplete"
          },
          "403": {
            "description": "The given Zip-Code doesn't exist or is not in the given city.\nBuyer with this Email already exists.\nInvalid Email.\nPassword is not secure."
          },
          "500": {
            "description": "Internal Server Error."
          },
          "201": {
            "description": "Buyer created",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "password": "Test123#",
                  "firstName": "Tarek",
                  "lastName": "Bürner",
                  "email": "test@test.de",
                  "iban": "DE1234567890",
                  "address": "Leiser Weg 12",
                  "zipCode": 70569,
                  "city": "Stuttgart",
                  "updated-at":"2024-01-07T17:15:24.623Z",
                  "created-at": "2024-01-07T17:15:24.623Z"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "buyer"
        ],
        "summary": "Get all buyers, not used in the frontend application",
        "operationId": "getBuyers",
        "responses": {
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Buyers returned",
            "content": {
              "application/json": {
                "example": [{
                  "_id": "cWGtlAnhcsVViidg",
                  "password": "Test123#",
                  "firstName": "Tarek",
                  "lastName": "Bürner",
                  "email": "test@test.de",
                  "iban": "DE1234567890",
                  "address": "Leiser Weg 12",
                  "zipCode": 70569,
                  "city": "Stuttgart",
                  "updated-at":"2024-01-07T17:15:24.623Z",
                  "created-at": "2024-01-07T17:15:24.623Z"
                },
                  {
                    "_id": "cWGtlAnhcsVViidk",
                    "password": "Test123#",
                    "firstName": "Tarek",
                    "lastName": "Bürner",
                    "email": "test@test.de",
                    "iban": "DE1234567890",
                    "address": "Leiser Weg 12",
                    "zipCode": 70569,
                    "city": "Stuttgart",
                    "updated-at":"2024-01-07T17:15:24.623Z",
                    "created-at": "2024-01-07T17:15:24.623Z"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/buyer/{buyerID}": {
      "get": {
        "tags": [
          "buyer"
        ],
        "summary": "Get Buyer by BuyerID",
        "operationId": "getBuyerByID",
        "responses": {
          "404": {
            "description": "Buyer not found"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Buyer returned",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "password": "Test123#",
                  "firstName": "Tarek",
                  "lastName": "Bürner",
                  "email": "test@test.de",
                  "iban": "DE1234567890",
                  "address": "Leiser Weg 12",
                  "zipCode": 70569,
                  "city": "Stuttgart",
                  "updated-at":"2024-01-07T17:15:24.623Z",
                  "created-at": "2024-01-07T17:15:24.623Z"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "buyer"
        ],
        "summary": "Update Buyer",
        "operationId": "updateBuyerByID",
        "requestBody": {
          "$ref": "#/components/requestBodies/BuyerPatch"
        },
        "responses": {
          "400": {
            "description": "User Data incomplete"
          },
          "403": {
            "description": "The given Zip-Code doesn't exist or is not in the given city.\nInvalid Email.\nPassword is not secure."
          },
          "401": {
            "description": "Wrong Password"
          },
          "404": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Buyer updated",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "password": "Test123#",
                  "firstName": "Tarek",
                  "lastName": "Bürner",
                  "email": "test@test.de",
                  "iban": "DE1234567890",
                  "address": "Leiser Weg 12",
                  "zipCode": 70569,
                  "city": "Stuttgart",
                  "updated-at":"2024-01-07T17:15:24.623Z",
                  "created-at": "2024-01-07T17:15:24.623Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "buyer"
        ],
        "summary": "Delete Buyer",
        "operationId": "deleteBuyer",
        "responses": {
          "401": {
            "description": "Wrong Password"
          },
          "404": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Buyer deleted"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
...
    "/article": {
      "post": {
        "tags": [
          "article"
        ],
        "summary": "Create a new article",
        "operationId": "createArticle",
        "requestBody": {
          "$ref": "#/components/requestBodies/Article"
        },
        "responses": {
          "400": {
            "description": "Article Data incomplete"
          },
          "401": {
            "description": "Authorization Header missing.\n Wrong Password"
          },
          "403": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "201": {
            "description": "Article created",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "title": "Xiaomi 13T Pro",
                  "description": "A really amazing Smartphone",
                  "price": 790.01,
                  "stockQuantity": 1000,
                  "visible": true,
                  "brand": "Xiaomi",
                  "seller": "cWGtlAnhcsVViidg",
                  "searchingKeywords": ["Smartphone", "Xiaomi", "keyword3"],
                  "createdAt":"2023-12-15T08:05:36.701Z",
                  "updatedAt":"2024-01-07T17:15:24.623Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "article"
        ],
        "summary": "Get all articles",
        "operationId": "getArticles",
        "responses": {
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Articles returned",
            "content": {
              "application/json": {
                "example": [{
                  "_id": "cWGtlAnhcsVViidg",
                  "title": "Xiaomi 13T Pro",
                  "description": "A really amazing Smartphone",
                  "price": 790.01,
                  "stockQuantity": 1000,
                  "visible": true,
                  "brand": "Xiaomi",
                  "seller": "cWGtlAnhcsVViidg",
                  "searchingKeywords": ["Smartphone", "Xiaomi", "keyword3"],
                  "createdAt":"2023-12-15T08:05:36.701Z",
                  "updatedAt":"2024-01-07T17:15:24.623Z"
                },
                  {
                    "_id": "cWGtlAnhcsVViidk",
                    "title": "Xiaomi 13T Pro",
                    "description": "A really amazing Smartphone",
                    "price": 790.01,
                    "stockQuantity": 1000,
                    "visible": true,
                    "brand": "Xiaomi",
                    "seller": "cWGtlAnhcsVViidg",
                    "searchingKeywords": ["Smartphone", "Xiaomi", "keyword3"],
                    "createdAt":"2023-12-15T08:05:36.701Z",
                    "updatedAt":"2024-01-07T17:15:24.623Z"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/article/{articleID}": {
      "get": {
        "tags": [
          "article"
        ],
        "summary": "Get Article by ArticleID",
        "operationId": "getArticleByID",
        "responses": {
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Article returned",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "title": "Xiaomi 13T Pro",
                  "description": "A really amazing Smartphone",
                  "price": 790.01,
                  "stockQuantity": 1000,
                  "visible": true,
                  "brand": "Xiaomi",
                  "seller": "cWGtlAnhcsVViidg",
                  "searchingKeywords": ["Smartphone", "Xiaomi", "keyword3"],
                  "createdAt":"2023-12-15T08:05:36.701Z",
                  "updatedAt":"2024-01-07T17:15:24.623Z"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "article"
        ],
        "summary": "Update Article",
        "operationId": "updateArticleByID",
        "requestBody": {
          "$ref": "#/components/requestBodies/ArticlePatch"
        },
        "responses": {
          "401": {
            "description": "Authorization Header missing.\n Wrong Password"
          },
          "403": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Seller updated",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "title": "Xiaomi 13T Pro",
                  "description": "A really amazing Smartphone",
                  "price": 790.01,
                  "stockQuantity": 1000,
                  "visible": true,
                  "brand": "Xiaomi",
                  "seller": "cWGtlAnhcsVViidg",
                  "searchingKeywords": ["Smartphone", "Xiaomi", "keyword3"],
                  "createdAt":"2023-12-15T08:05:36.701Z",
                  "updatedAt":"2024-01-07T17:15:24.623Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "article"
        ],
        "summary": "Delete Article",
        "operationId": "deleteArticle",
        "responses": {
          "401": {
            "description": "Authorization Header missing.\n Wrong Password"
          },
          "403": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Article deleted"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Article": {
        "type": "object",
        "required": [
          "title",
          "description",
          "price",
          "stockQuantity",
          "visible",
          "brand",
          "searchingKeywords"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "example": "Xiaomi 13T Pro"
          },
          "description": {
            "type": "string",
            "example": "A really amazing Smartphone"
          },
          "price": {
            "type": "number",
            "format": "double",
            "multipleOf": 0.01,
            "example": 790.01
          },
          "stockQuantity": {
            "type": "integer",
            "example": 1000
          },
          "visible": {
            "type": "boolean",
            "example": true
          },
          "brand": {
            "type": "string",
            "example": "Xiaomi"
          },
          "searchingKeywords": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Smartphone", "Xiaomi", "keyword3"]
          },
          "seller": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          }
        }
      },
      "ArticlePatch": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "example": "Xiaomi 13T Pro"
          },
          "description": {
            "type": "string",
            "example": "A really amazing Smartphone"
          },
          "price": {
            "type": "number",
            "format": "double",
            "multipleOf": 0.01,
            "example": 790.01
          },
          "stockQuantity": {
            "type": "integer",
            "example": 1000
          },
          "visible": {
            "type": "boolean",
            "example": true
          },
          "brand": {
            "type": "string",
            "example": "Xiaomi"
          },
          "searchingKeywords": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Smartphone", "Xiaomi", "keyword3"]
          }
        }
      },
...
      "BuyerPatch": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "password": {
            "type": "string",
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=!]).{8,}$",
            "example": "Test123#"
          },
          "firstName": {
            "type": "string",
            "example": "Tarek"
          },
          "lastName": {
            "type": "string",
            "example": "Bürner"
          },
          "email": {
            "type": "string",
            "example": "test@test.de"
          },
          "address": {
            "type": "string",
            "example": "Leiser Weg 12"
          },
          "zipCode": {
            "type": "number",
            "pattern": "5 digits, existing zipCode",
            "example": 70569
          },
          "city": {
            "type": "string",
            "pattern": "must match the zipCode",
            "example": "Stuttgart"
          }
        }
      },
      "Buyer": {
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "iban",
          "address",
          "zipCode",
          "city"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "password": {
            "type": "string",
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=!]).{8,}$",
            "example": "Test123#"
          },
          "firstName": {
            "type": "string",
            "example": "Tarek"
          },
          "lastName": {
            "type": "string",
            "example": "Bürner"
          },
          "email": {
            "type": "string",
            "example": "test@test.de"
          },
          "address": {
            "type": "string",
            "example": "Leiser Weg 12"
          },
          "zipCode": {
            "type": "number",
            "pattern": "5 digits, existing zipCode",
            "example": 70569
          },
          "city": {
            "type": "string",
            "pattern": "must match the zipCode",
            "example": "Stuttgart"
          }
        }
      },
...
    },
    "requestBodies": {
      "Buyer": {
        "description": "Buyer, that should be registered",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Buyer"
            }
          }
        },
        "required": true
      },
      "BuyerPatch": {
        "description": "Buyer, that should be updated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BuyerPatch"
            }
          }
        }
      },
...
      "Article": {
        "description": "Seller, that should be updated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Article"
            }
          }
        }
      },
      "ArticlePatch": {
        "description": "Seller, that should be updated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArticlePatch"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
...
  }
}